+++ core/wallet-store-inmemory/src/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './StoreInternal.js'
+export * from './store-internal.js'
